requirements.txt
asgiref==3.7.2
Django==5.0.3
django-extra-views==0.14.0
django-filter==24.1
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
